Refactor tests to improve organization and clarity.
