+++ solo-leveling-player-service/src/test/resources/testcontainers.properties
@@ -1,0 +1,2 @@
+checks.disable=true
+testcontainers.reuse.enable=true
